Us 1 tests

+++ src/main/java/com/mercadolibre/group8_bootcamp_finalproject/dtos/response/InboundOrderResponseDTO.java
@@ -25,7 +25,7 @@ public class InboundOrderResponseDTO {
             BatchResponseDTO batchResponseDTO = BatchResponseDTO.builder()
                     .batchId(batch.getId())
                     .batchNumber(batch.getNumber())
-                    .productId(batch.getProduct().getId().intValue())
+                    .productId(batch.getProduct().getId())
                     .currentTemperature(batch.getCurrentTemperature())
                     .minimumTemperature(batch.getProduct().getMinTemperature())
                     .quantity(batch.getQuantity())
